Correct append to prepend
This is based off of the description in the documentation.
Since this correction isnt based off of the source code
(sorry, trying this PR out on gitpod.io since I'm away from my computer),
it is possible that this correction should be reversed.

This is also a very small PR, I am happy if it goes in
with another, more meaningful contribution.