Update release config with cargo-dist@v0.0.4
This is generated using cargo-dist@v0.0.4 and running:

cargo dist init --ci=github --installer=shell --installer=powershell

This change introduces a new target, namely `aarch64-apple-darwin`, and
also generates installer scripts for shell and powershell.
Both of these additions can be turned off if needed.

Hopefully, this change might fix some of the issues with cargo-binstall
not finding the binary.

The toolchain used for building in release is also fixed to 1.57.1, same
as it was in the prior version of the release CI.

I didn't notice any custom configuration made to the files that should
be ported over, but I might have missed some.