Add support for fixed length arrays
These have been a problem in a couple structures I'm hoping to translate
into Typescript.

I'm unsure of the syntax for all the languages excepting typescript. For
the majority of them I reused Vec (because a fixed length array is a
specialised version of a dynamically sized array).

For go, I found something that suggested this syntax `[4]string`. I
don't know if this is right, so I'd appreciate input on the PR.

I would appreciate input in the PR for go, kotlin, scala and swift.