Fix writing enum values in Go (#233)
For Algebraic Rust Enums, when generating them in Go, the values of the constants were the original name of the enum variant instead of the renamed one, which takes into consideration the serde attributes. Because of that, it could happen that in such enums the unmarshaling of JSON will fail due to not having any content to process. That was caused by the match case not hitting any variant.